Extract MemberClaimsView and MemberGeneralView from MemberInformation, add MemberCard to simplify overview
