+++ src/features/claims/claim-details/MemberInformation/components/MemberClaimsView.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const MemberClaimsView: React.FC<{}> = () => {
+  return null
+}
