+++ src/features/commands/use-command-line.tsx
@@ -115,7 +115,6 @@ export const CommandLineProvider: React.FC = ({ children }) => {
     })
 
     if (matchIndex > -1) {
-      TagManager.dataLayer({ dataLayer: { event: 'shortcut_used' } })
       actions.current[matchIndex].onResolve()
       setShowCommandLine(false)
     }
