+++ src/pages/claims/ClaimDetailsPage.tsx
@@ -22,7 +22,7 @@ import { ClaimReserve } from 'features/claims/claim-details/ClaimReserve'
 import { ClaimRestrictionInformation } from 'features/claims/claim-details/ClaimRestrictionInformation'
 import { ClaimTranscriptions } from 'features/claims/claim-details/ClaimTranscriptions'
 import { ClaimType } from 'features/claims/claim-details/ClaimType/ClaimType'
-import { MemberInformation } from 'features/claims/claim-details/MemberInformation'
+import { MemberInformation } from 'features/claims/claim-details/MemberInformation/MemberInformation'
 import { useCommandLine } from 'features/commands/use-command-line'
 import { ChatPane } from 'features/member/tabs/ChatPane'
 import { getCarrierText } from 'features/member/tabs/contracts-tab/utils'
